+++ CHANGELOG.md
@@ -5,6 +5,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- parsing the payments in-lieu of dividends
+
+### Removed
+
+- 
+
 ## [1.1.0] - 2023-03-10
 
 ### Added